+++ src/translations/en/jitter.php
@@ -1,20 +1,15 @@
 <?php
 /**
- * Jitter plugin for Craft CMS 3.x
+ * Jitter plugin for Craft CMS 4.x
  *
  * A just in time image transformation service.
  *
  * @link      https://kyleandrews.dev/
- * @copyright Copyright (c) 2020 Kyle Andrews
+ * @copyright Copyright (c) 2022 Kyle Andrews
  */
 
 /**
  * Jitter en Translation
- *
- * Returns an array with the string to be translated (as passed to `Craft::t('jitter', '...')`) as
- * the key, and the translation as the value.
- *
- * http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html
  *
  * @author    Kyle Andrews
  * @package   Jitter